requirements.txt
Flask==2.3.2
google-generativeai==0.8.5
python-dotenv==1.1.1
gunicorn==23.0.0
wikipedia
